requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
black==26.1.0
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contextlib2==21.6.0
cryptography==46.0.3
distlib==0.4.0
dj-database-url==3.1.0
Django==5.2
django-cors-headers==4.9.0
django-ninja==1.5.1
django-ninja-extra==0.30.8
django-ninja-jwt==5.4.2
dnspython==2.8.0
email-validator==2.3.0
filelock==3.20.3
gunicorn==23.0.0
h11==0.16.0
identify==2.6.16
idna==3.11
iniconfig==2.3.0
injector==0.23.0
mailtrap==2.4.0
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==25.0
pathspec==1.0.3
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.1
psycopg==3.3.2
psycopg-binary==3.3.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-django==4.11.1
python-decouple==3.8
python-dotenv==1.2.1
pytokens==0.4.0
PyYAML==6.0.3
requests==2.32.5
ruff==0.14.13
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
virtualenv==20.36.1
whitenoise==6.11.0
